Fix some warnings in clang 3.7.1

+++ include/chaiscript/dispatchkit/dispatchkit.hpp
@@ -1446,7 +1446,7 @@ namespace chaiscript
         static typename Container::const_iterator find_keyed_value(const Container &t_c, const Key &t_key, const size_t t_hint)
           {
             if (t_c.size() > t_hint && t_c[t_hint].first == t_key) {
-              return t_c.begin() + t_hint;
+              return t_c.begin() + static_cast<long>(t_hint);
             } else {
               return find_keyed_value(t_c, t_key);
             }
